+++ src/app/map/map.component.spec.ts
@@ -1,4 +1,7 @@
-import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
+import {
+    BrowserAnimationsModule,
+    NoopAnimationsModule,
+} from '@angular/platform-browser/animations';
 import { HttpClientTestingModule } from '@angular/common/http/testing';
 import { ComponentFixture, TestBed } from '@angular/core/testing';
 import { CurrentUserService } from '@services/current-user.service';
@@ -55,6 +58,7 @@ describe('MapComponent', () => {
             declarations: [MapComponent],
             imports: [
                 BrowserAnimationsModule,
+                NoopAnimationsModule,
                 MatAutocompleteModule,
                 HttpClientTestingModule,
                 FormsModule,