+++ src/app/map/map.component.ts
@@ -90,7 +90,7 @@ export class MapComponent implements OnInit {
     currentEvent: number; //change to subject?
     currentEventName: string;
     sitesDataArray: any;
-    eventMarkers = L.featureGroup([]);
+    siteMarkers = L.featureGroup([]);
 
     mapScale;
     latitude;
@@ -137,6 +137,7 @@ export class MapComponent implements OnInit {
     states$: Observable<State[]>;
 
     //These variables indicate if each layer is checked
+    sitesVisible = true;
     watershedsVisible = false;
     currWarningsVisible = false;
     watchWarnVisible = false;
@@ -290,7 +291,7 @@ export class MapComponent implements OnInit {
             this.mapResults(
                 this.allSites,
                 this.siteIcon,
-                this.siteService.siteMarkers,
+                this.siteService.allSiteMarkers,
                 false
             );
         });
@@ -395,12 +396,9 @@ export class MapComponent implements OnInit {
             this.currentEventName = this.events[0].event_name;
         }
 
-        //Clear the old event markers from the map
-        if (this.eventMarkers !== undefined) {
-            this.eventMarkers.removeFrom(this.map);
-        }
         //Clear the old markers from the layer
-        this.eventMarkers = L.featureGroup([]);
+        this.siteService.siteMarkers.clearLayers();
+
         //Plot markers for selected event
         this.siteService
             .getEventSites(this.currentEvent)
@@ -410,7 +408,7 @@ export class MapComponent implements OnInit {
                 this.mapResults(
                     this.sitesDataArray,
                     this.eventIcon,
-                    this.eventMarkers,
+                    this.siteService.siteMarkers,
                     true
                 );
                 setTimeout(() => {
@@ -430,8 +428,9 @@ export class MapComponent implements OnInit {
         });
 
         this.supplementaryLayers = {
+            Sites: this.siteService.siteMarkers,
             Watersheds: MAP_CONSTANTS.mapLayers.esriDynamicLayers.HUC,
-            'All STN Sites': this.siteService.siteMarkers,
+            'All STN Sites': this.siteService.allSiteMarkers,
             'Current Warnings*':
                 MAP_CONSTANTS.mapLayers.esriFeatureLayers.currentWarnings,
             'Watches/Warnings*':
@@ -507,6 +506,9 @@ export class MapComponent implements OnInit {
         // When layer is checked, add layer icon to legend
         /* istanbul ignore next */
         this.map.on('overlayadd', (e) => {
+            if (e.name === 'Sites') {
+                this.sitesVisible = true;
+            }
             if (e.name === 'Watersheds') {
                 this.watershedsVisible = true;
             }
@@ -529,6 +531,9 @@ export class MapComponent implements OnInit {
         // When layer is unchecked, remove layer icon from legend
         /* istanbul ignore next */
         this.map.on('overlayremove', (e) => {
+            if (e.name === 'Sites') {
+                this.sitesVisible = false;
+            }
             if (e.name === 'Watersheds') {
                 this.watershedsVisible = false;
             }
@@ -558,10 +563,10 @@ export class MapComponent implements OnInit {
             this.currentZoom = this.map.getZoom();
             //Disable clustering for the All STN Sites layer when zoom >= 12 so we can see individual sites
             if (this.currentZoom >= 12) {
-                this.siteService.siteMarkers.disableClustering();
+                this.siteService.allSiteMarkers.disableClustering();
             }
             if (this.currentZoom < 12) {
-                this.siteService.siteMarkers.enableClustering();
+                this.siteService.allSiteMarkers.enableClustering();
             }
             //If the zoom went from 9 to 8 and the gages/watches/warnings are on,
             //that layer is checked, but it's not displayed
@@ -716,6 +721,9 @@ export class MapComponent implements OnInit {
         // When layer is checked, add layer icon to legend
         /* istanbul ignore next */
         this.map.on('overlayadd', (e) => {
+            if (e.name === 'Sites') {
+                this.sitesVisible = true;
+            }
             if (e.name === 'Watersheds') {
                 this.watershedsVisible = true;
             }
@@ -738,6 +746,9 @@ export class MapComponent implements OnInit {
         // When layer is unchecked, remove layer icon from legend
         /* istanbul ignore next */
         this.map.on('overlayremove', (e) => {
+            if (e.name === 'Sites') {
+                this.sitesVisible = false;
+            }
             if (e.name === 'Watersheds') {
                 this.watershedsVisible = false;
             }
@@ -762,8 +773,8 @@ export class MapComponent implements OnInit {
     eventFocus() {
         //If there are site markers, zoom to those
         //Otherwise, zoom back to default extent
-        if (this.map.hasLayer(this.eventMarkers)) {
-            this.map.fitBounds(this.eventMarkers.getBounds());
+        if (this.map.hasLayer(this.siteService.siteMarkers)) {
+            this.map.fitBounds(this.siteService.siteMarkers.getBounds());
         } else {
             this.map.setView(
                 MAP_CONSTANTS.defaultCenter,
@@ -861,14 +872,14 @@ export class MapComponent implements OnInit {
                         site.site_no !== 'ASTUT27853' &&
                         site.site_no !== 'AZGRA27856'
                     ) {
-                        //put all the event markers in the same layer group
-                        if (layerType == this.eventMarkers) {
+                        //put all the site markers in the same layer group
+                        if (layerType == this.siteService.siteMarkers) {
                             L.marker([lat, long], { icon: myIcon })
                                 .bindPopup(popupContent)
                                 .addTo(layerType);
                         }
                         //Make circle markers for the All STN Sites layer
-                        if (layerType == this.siteService.siteMarkers) {
+                        if (layerType == this.siteService.allSiteMarkers) {
                             L.marker([lat, long], {
                                 icon: myIcon,
                                 iconSize: 32,
@@ -880,8 +891,8 @@ export class MapComponent implements OnInit {
                 }
             }
         }
-        if (layerType == this.eventMarkers) {
-            this.eventMarkers.addTo(this.map);
+        if (layerType == this.siteService.siteMarkers) {
+            this.siteService.siteMarkers.addTo(this.map);
             //When filtering sites, zoom to layer, and open map pane
             if (zoomToLayer == true) {
                 this.eventFocus();
@@ -968,9 +979,8 @@ export class MapComponent implements OnInit {
                 bracketTrue;
 
             //Clear current markers when a new filter is submitted
-            if (this.map.hasLayer(this.eventMarkers)) {
-                this.eventMarkers.removeFrom(this.map);
-                this.eventMarkers = L.featureGroup([]);
+            if (this.map.hasLayer(this.siteService.siteMarkers)) {
+                this.siteService.siteMarkers.clearLayers();
             }
             //Find sites that match the user's query
             this.siteService
@@ -988,7 +998,7 @@ export class MapComponent implements OnInit {
                         this.mapResults(
                             res,
                             this.eventIcon,
-                            this.eventMarkers,
+                            this.siteService.siteMarkers,
                             true
                         );